molecule_loader_api: Added extra 'arg' parameter which is required for LAMMPS data format. fix bug in chain assignment. fixed crash issue in spatial_hash where cells were overpopulated by incorrect length units or malformed datasets.

+++ src/core/md_spatial_hash.h
@@ -95,8 +95,8 @@ static inline vec3_t md_spatial_acc_iter_pos(const md_spatial_acc_iter_t* iter) 
 
 // Initialize a spatial hash structure with a set of coordinates given by array of vec3_t (xyz) or separate arrays (x,y,z).
 // pbc_ext is optional and supplies the periodic extent for each axis. To disable periodicity, supply (0,0,0).
-md_spatial_hash_t* md_spatial_hash_create_vec3(const vec3_t in_xyz[], const int32_t in_idx[], int64_t count, const struct md_unit_cell_t* unit_cell, struct md_allocator_i* alloc);
-md_spatial_hash_t* md_spatial_hash_create_soa (const float in_x[], const float in_y[], const float in_z[], const int32_t in_idx[], int64_t count, const struct md_unit_cell_t* unit_cell, struct md_allocator_i* alloc);
+md_spatial_hash_t* md_spatial_hash_create_vec3(const vec3_t in_xyz[], const int32_t in_idx[], size_t count, const struct md_unit_cell_t* unit_cell, struct md_allocator_i* alloc);
+md_spatial_hash_t* md_spatial_hash_create_soa (const float in_x[], const float in_y[], const float in_z[], const int32_t in_idx[], size_t count, const struct md_unit_cell_t* unit_cell, struct md_allocator_i* alloc);
 
 void md_spatial_hash_free(md_spatial_hash_t* spatial_hash);
 
@@ -108,7 +108,7 @@ void md_spatial_hash_query_batch(const md_spatial_hash_t* spatial_hash, vec3_t p
 
 // Get a list of indices which fall within the search space (pos + radius)
 // Writes directly to the supplied buffer and will return the number of indices written.
-int64_t md_spatial_hash_query_idx(int32_t* buf_idx, int64_t buf_cap, const md_spatial_hash_t* spatial_hash, vec3_t pos, float radius);
+size_t  md_spatial_hash_query_idx(int32_t* buf_idx, size_t buf_cap, const md_spatial_hash_t* spatial_hash, vec3_t pos, float radius);
 void    md_spatial_hash_query_bits(struct md_bitfield_t* bf, const md_spatial_hash_t* spatial_hash, vec3_t pos, float radius);
 
 #ifdef __cplusplus
